Update code to work with ws@1.x
Since ws removed client code we had to make a special export
just for use in browser land.  This is a breaking change
