+++ package.json
@@ -28,7 +28,7 @@
     "debug": "^2.2.0",
     "lodash.bind": "^3.1.0",
     "uuid": "^2.0.1",
-    "ws": "^1.0.0"
+    "ws": "^1.0.1"
   },
   "pre-commit": [
     "lint",
